Fix scroll clamping at u16::MAX (65535 lines)
The scroll position ran through viewport_offset(), which clamped the
result to u16::MAX, so a buffer with more than 65535 lines got stuck near
line 65535 in follow-tail mode — new messages past that never showed (e.g.
with SIM_CLI_LOG_BUFFER / DEFAULT_MAX raised to 500000, the view froze
around 65578).

Since rendering is virtualized (only the visible window is built), the
u16 ceiling is unnecessary: viewport_offset now returns u64 and the
renderer selects the window in u64, feeding ratatui only a small intra-line
offset. RenderOutput.total_lines and render_ratatui's return become u64
too, and the "N new" hint no longer truncates to u16.

Adds regression tests: viewport_offset past u16::MAX, and a 70k-line buffer
following to the real tail.

+++ src/frontend.rs
@@ -302,7 +302,7 @@ impl Frontend {
 
     /// 每帧渲染后：记录视口尺寸并刷新未读行数(驱动 "▼ N new" 提示)。
     fn apply_output(&mut self, out: &crate::ui::render_state::RenderOutput) {
-        self.scrollback.on_frame(out.viewport_height, out.total_lines as u64);
+        self.scrollback.on_frame(out.viewport_height, out.total_lines);
     }
 
     pub async fn run<B: Backend>(&mut self, term: &mut Terminal<B>) -> Result<()> {
